fronti1.2/requirements.txt
Flask
flask-cors
torch
transformers
Pillow
openai
gunicorn